% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ref_rows.R
\name{add_ref_rows}
\alias{add_ref_rows}
\title{Title}
\usage{
add_ref_rows(tidy_model, data, delim = "")
}
\arguments{
\item{tidy_model}{A model that has been tidied using \link[broom:tidy]{broom::tidy}}

\item{data}{the dataframe used to fit \code{tidy_model}}

\item{delim}{a character value indicating how factor levels are
appended to factor variable names.}
}
\value{
a \link[tibble:tibble-package]{tibble} with additional rows
representing the reference category.
}
\description{
Title
}
\examples{



}
