% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_perc.R, R/point_error.R, R/point_gap.R
\name{as_numPer}
\alias{as_numPer}
\alias{as_pointErr}
\alias{as_pointGap}
\title{tblString casting}
\usage{
as_numPer(x)

as_pointErr(x)

as_pointGap(x)
}
\arguments{
\item{x}{object to be coerced. Eligible object types are
\itemize{
\item \code{list}
\item \code{matrix}
\item \code{data.frame}
}

The values of \code{x} should be ordered in the same manner as the
inputs of the \code{tblString} type. For example, coercing a list
into a \code{pointGap} object would require the list to have three
elements in the order of point, lower, and upper values.}
}
\value{
a \code{tblString} vector.
}
\description{
Coerces object(s) into \code{tblString} types.
}
\note{
data in \code{x} must be numeric.
}
\examples{
as_numPer(list(1,2))
as_numPer(list(c(1:10), 100))
as_pointErr(list(1,2))
as_pointErr(list(c(1:10), (1:10)*10))

x <- list(c(1,2,3), c(0,1,2), c(2,3,4))

as_pointGap(x)

x <- as.data.frame(x)

as_pointGap(x)

x <- c(1L, 0L, 2L)

as_pointGap(x)

x <- matrix(c(1,2,3,0,1,2,2,3,4), ncol=3)

as_pointGap(x)


}
