% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_perc.R, R/point_error.R, R/point_gap.R
\name{format.tblStrings_numPer}
\alias{format.tblStrings_numPer}
\alias{format.tblStrings_pointErr}
\alias{format.tblStrings_pointGap}
\title{tblString formats}
\usage{
\method{format}{tblStrings_numPer}(x, ...)

\method{format}{tblStrings_pointErr}(x, ...)

\method{format}{tblStrings_pointGap}(x, ...)
}
\arguments{
\item{x}{an object to print}

\item{...}{not currently used}
}
\value{
printed output in the console window
}
\description{
\code{tblString}objects are printed with some guiding principles.
\itemize{
\item table values should be rounded based on their magnitude,
not based on a uniform number of rounding digits.
\item data should be spaced appropriately, not smushed together.
}
}
\section{Functions}{
\itemize{
\item \code{format.tblStrings_numPer}: 

\item \code{format.tblStrings_pointErr}: 
}}

\examples{

pointGap(1, 0, 2, max_decimal = 0)
pointGap(1, 0, 2, max_decimal = 1)
pointGap(1, 0, 2, max_decimal = 2)

pointGap(1e4, 1e3, 2e4)
pointGap(1e4, 1e3, 2e4, big_mark = '-')
pointGap(1e4, 1e3, 2e4, brac_left = '[', brac_right = ']', brac_middle = '-')

}
