% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_gap.R
\name{is_pointGap}
\alias{is_pointGap}
\title{Point gaps}
\usage{
is_pointGap(x)
}
\arguments{
\item{x}{an object}
}
\value{
a logical vector of length equal to \code{x}
}
\description{
Point gaps
}
\examples{

is_pointGap(1)

is_pointGap(pointGap(1,0,2))


}
