% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{is_pval}
\alias{is_pval}
\title{P-values}
\usage{
is_pval(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\value{
\code{TRUE} if \code{x} inherits from the \code{pval} class, \code{FALSE} otherwise
}
\description{
P-values
}
\examples{

is_pval(1/2)

is_pval(pval(1/2))

}
