% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_perc.R
\name{np_perc_gt}
\alias{np_perc_gt}
\alias{np_nums_gt}
\alias{np_perc_gteq}
\alias{np_nums_gteq}
\alias{np_perc_lt}
\alias{np_nums_lt}
\alias{np_perc_lteq}
\alias{np_nums_lteq}
\title{numPer comparisons}
\usage{
np_perc_gt(x, value)

np_nums_gt(x, value)

np_perc_gteq(x, value)

np_nums_gteq(x, value)

np_perc_lt(x, value)

np_nums_lt(x, value)

np_perc_lteq(x, value)

np_nums_lteq(x, value)
}
\arguments{
\item{x}{a \code{numPer} vector}

\item{value}{a numeric value to compare with the \code{numPer} vector's data.}
}
\value{
a logical vector of length equal to the size of \code{x}
}
\description{
Compare the counts and percentages comprised by a \code{numPer}
object to standard values. The names of these functions indicate their
behavior. Each functions starts with \code{np} to indicate that it is
specifically for \code{numPer} objects. The middle term indicates which
part of the \code{numPer} is compared to \code{value}.
\itemize{
\item \code{perc}: compare the \code{numPer}'s percent to \code{value}
\item \code{nums}: compare the \code{numPer}'s counts to \code{value}
}

The last term indicates the comparison operator.
\itemize{
\item \code{gt}: strictly greater than
\item \code{gteq}: greater than or equal to
\item \code{lt}: strictly less than
\item \code{lteq}: less than or equal to
}
}
\note{
the percent value for a \code{numPer} vector is between 0 and 1.
Comparisons between percents and \code{value} should be made accordingly.
}
\examples{

# 20 out of 100 - 20\%
np <- numPer(20, 100)

np_perc_gt(np, 1/5)
np_perc_lt(np, 1/5)
np_perc_gteq(np, 1/5)
np_perc_lteq(np, 1/5)


}
