% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_perc.R, R/point_error.R, R/point_gap.R,
%   R/pvalue.R
\name{obj_print_data.tblStrings_numPer}
\alias{obj_print_data.tblStrings_numPer}
\alias{vec_ptype_abbr.tblStrings_numPer}
\alias{vec_cast.tblStrings_numPer}
\alias{vec_ptype2.tblStrings_numPer}
\alias{vec_proxy_equal.tblStrings_numPer}
\alias{vec_proxy_compare.tblStrings_numPer}
\alias{vec_math.tblStrings_numPer}
\alias{vec_arith.tblStrings_numPer}
\alias{obj_print_data.tblStrings_pointErr}
\alias{vec_ptype_abbr.tblStrings_pointErr}
\alias{vec_cast.tblStrings_pointErr}
\alias{vec_ptype2.tblStrings_pointErr}
\alias{vec_proxy_equal.tblStrings_pointErr}
\alias{vec_proxy_compare.tblStrings_pointErr}
\alias{vec_math.tblStrings_pointErr}
\alias{vec_arith.tblStrings_pointErr}
\alias{obj_print_data.tblStrings_pointGap}
\alias{vec_ptype_abbr.tblStrings_pointGap}
\alias{vec_cast.tblStrings_pointGap}
\alias{vec_ptype2.tblStrings_pointGap}
\alias{vec_proxy_equal.tblStrings_pointGap}
\alias{vec_proxy_compare.tblStrings_pointGap}
\alias{vec_math.tblStrings_pointGap}
\alias{vec_arith.tblStrings_pointGap}
\alias{format.tblStrings_pval}
\alias{obj_print_data.tblStrings_pval}
\alias{vec_ptype_abbr.tblStrings_pval}
\alias{vec_ptype2.tblStrings_pval}
\alias{vec_cast.tblStrings_pval}
\alias{vec_proxy_equal.tblStrings_pval}
\alias{vec_proxy_compare.tblStrings_pval}
\alias{vec_math.tblStrings_pval}
\alias{vec_arith.tblStrings_pval}
\title{Do not use}
\usage{
\method{obj_print_data}{tblStrings_numPer}(x)

\method{vec_ptype_abbr}{tblStrings_numPer}(x, ...)

\method{vec_cast}{tblStrings_numPer}(x, to, ...)

\method{vec_ptype2}{tblStrings_numPer}(x, y, ...)

\method{vec_proxy_equal}{tblStrings_numPer}(x, ...)

\method{vec_proxy_compare}{tblStrings_numPer}(x, ...)

\method{vec_math}{tblStrings_numPer}(.fn, .x, ...)

\method{vec_arith}{tblStrings_numPer}(op, x, y, ...)

\method{obj_print_data}{tblStrings_pointErr}(x)

\method{vec_ptype_abbr}{tblStrings_pointErr}(x, ...)

\method{vec_cast}{tblStrings_pointErr}(x, to, ...)

\method{vec_ptype2}{tblStrings_pointErr}(x, y, ...)

\method{vec_proxy_equal}{tblStrings_pointErr}(x, ...)

\method{vec_proxy_compare}{tblStrings_pointErr}(x, ...)

\method{vec_math}{tblStrings_pointErr}(.fn, .x, ...)

\method{vec_arith}{tblStrings_pointErr}(op, x, y, ...)

\method{obj_print_data}{tblStrings_pointGap}(x)

\method{vec_ptype_abbr}{tblStrings_pointGap}(x, ...)

\method{vec_cast}{tblStrings_pointGap}(x, to, ...)

\method{vec_ptype2}{tblStrings_pointGap}(x, y, ...)

\method{vec_proxy_equal}{tblStrings_pointGap}(x, ...)

\method{vec_proxy_compare}{tblStrings_pointGap}(x, ...)

\method{vec_math}{tblStrings_pointGap}(.fn, .x, ...)

\method{vec_arith}{tblStrings_pointGap}(op, x, y, ...)

\method{format}{tblStrings_pval}(x, ...)

\method{obj_print_data}{tblStrings_pval}(x)

\method{vec_ptype_abbr}{tblStrings_pval}(x, ...)

\method{vec_ptype2}{tblStrings_pval}(x, y, ...)

\method{vec_cast}{tblStrings_pval}(x, to, ...)

\method{vec_proxy_equal}{tblStrings_pval}(x, ...)

\method{vec_proxy_compare}{tblStrings_pval}(x, ...)

\method{vec_math}{tblStrings_pval}(.fn, .x, ...)

\method{vec_arith}{tblStrings_pval}(op, x, y, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{not used.}

\item{to}{an object.}

\item{y}{an object.}

\item{.fn}{a function.}

\item{.x}{an object.}

\item{op}{an operation.}
}
\description{
these functions are exported in order to ensure consistency
with methods in the \code{vctrs} package and should not be used directly.
}
