% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_gap.R
\name{pg_flip}
\alias{pg_flip}
\title{Flip}
\usage{
pg_flip(x)
}
\arguments{
\item{x}{a point gap object (see \link{pointGap})}
}
\value{
a \code{pointGap} object with flipped coordinates.
}
\description{
Multiplying a point-gap by negative 1 is sometimes
useful, but \code{pointGap} objects won't let you multiply them like that.
Instead, use \code{pg_flip()} which will automatically switch the upper and
lower bounds of the gap so it pg_covers the same numbers multiplied
by negative 1.
}
\examples{

ptr <- pointGap(2, 1, 4)

# negate
pg_flip(ptr)

# double negate = no change
pg_flip(pg_flip(ptr))


}
