% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_gap.R
\name{pg_misRef}
\alias{pg_misRef}
\title{Reference labels}
\usage{
pg_misRef(x)
}
\arguments{
\item{x}{a point gap object (see \link{pointGap})}
}
\value{
a \code{pointGap} object with reference values replacing values
with all NA.
}
\description{
Reference labels
}
