% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_gap.R
\name{pg_set_ref}
\alias{pg_set_ref}
\title{Reference labels}
\usage{
pg_set_ref(x, index)
}
\arguments{
\item{x}{a point gap object (see \link{pointGap})}

\item{index}{a logical or numeric vector indicating which values
should be set as reference.}
}
\value{
a \code{pointGap} object with reference values replacing values
with all NA.
}
\description{
Reference labels
}
\examples{

p <- pointGap(c(NA, 2), c(NA, 1.8), c(NA, 2.2), ref_value = 1)
pg_set_ref(p, index = 1)

}
