% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{pval}
\alias{pval}
\title{P-value}
\usage{
pval(
  x = double(),
  print_thresh = 0.001,
  column_label = "P-value",
  alpha_level = 0.05
)
}
\arguments{
\item{x}{a numeric value or vector with values greater than 0
and less than 1.}

\item{print_thresh}{a numeric value indicating the lowest number that
should be displayed. Any numbers below this threshold will be
written as < threshold in the output.}

\item{column_label}{a character value that could be used as a
column header for the p-value in a published table.}

\item{alpha_level}{a numeric value indicating the nominal type
1 error rate for your analysis.}
}
\value{
a \code{pval} object
}
\description{
P-values are created based on probability inputs.
}
\note{
the number of decimal digits to display is determined based
on the threshold value. For example, if you set threshold as 0.01,
then only 2 digits are shown. Similarly, a threshold of 0.001 will
lead to displaying 3 digits.
}
\examples{

pval(1/2)
pval(0.0001)
pval(c(0.1, 0.01, 0.001, 0.0001))
pval(c(0.1, 0.01, 0.001, 0.0001), print_thresh = 0.01)

}
