% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_p.R
\name{tbl_p}
\alias{tbl_p}
\title{P-value rounding}
\usage{
tbl_p(x, decimals = 3, thresh_lo = 0.001, thresh_hi = 0.999)
}
\arguments{
\item{x}{a vector of numeric values}

\item{decimals}{number of decimals to print}

\item{thresh_lo}{the lowest value printed. Values lower than the
threshold will be printed as <threshold.}

\item{thresh_hi}{the highest value printed. Values higher than the
threshold will be printed as >threshold.}
}
\value{
a character vector
}
\description{
P-value rounding
}
\examples{

pvals <- c(0.01, 0.001, 0.0001, 0.9, 0.99, 0.9999)
tbl_p(pvals)

}
