% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_pint.R
\name{tbl_pint}
\alias{tbl_pint}
\title{Point and interval (pint)}
\usage{
tbl_pint(
  estimate,
  lower,
  upper,
  bracket_left = "(",
  bracket_right = ")",
  bracket_sep = ", ",
  ...
)
}
\arguments{
\item{estimate}{point estimate values}

\item{lower}{lower values of interval estimate}

\item{upper}{upper values of interval estimate}

\item{bracket_left}{character value that will close the interval on the left}

\item{bracket_right}{character value that will close the interval on the right}

\item{bracket_sep}{character value that will separate the lower and upper interval estimates}

\item{...}{
  Arguments passed on to \code{\link[=tbl_string]{tbl_string}}
  \describe{
    \item{\code{.sep}}{Separator used to separate elements}
    \item{\code{.envir}}{environment to evaluate each expression in.}
    \item{\code{round_half_to_even}}{a logical value. If \code{TRUE}, then when the
fractional part of \code{x} is half-way between two rounding boundaries,
it will be rounded to the even boundary nearest to x. For example,
+23.5 becomes +24, as does +24.5; while −23.5 becomes −24,
as does −24.5. If \code{FALSE}, halves are automatically rounded to
the boundary with highest absolute value.}
    \item{\code{breaks}}{a positive, monotonically increasing numeric vector
designating rounding boundaries. With \code{breaks = c(1, 10, 100, 1000)}
and \code{decimals = c(2, 1, 0)}}
    \item{\code{decimals}}{a numeric vector of equal length to \code{breaks} that
indicates how many decimals to round to in the numeric range
designated by \code{breaks}. (see notes for example). Allowed values
are 0 <= \code{decimals} <= 20.}
    \item{\code{miss_replace}}{a character string that replaces missing values.}
    \item{\code{big_mark}}{a character value used to separate number groups to the
left of the decimal point. See \link{prettyNum} for more details on this.
Set this input to '' to negate it's effect.}
    \item{\code{big_interval}}{a numeric value indicating the size of number groups
for numbers before (hence big) the decimal.}
    \item{\code{small_mark}}{a character value used to separate number groups to
the right of the decimal point.}
    \item{\code{small_interval}}{a numeric value indicating the size of number groups
for numbers after (hence small) the decimal.}
    \item{\code{decimal_mark}}{the character to be used to indicate the numeric
decimal point.}
    \item{\code{zero_print}}{a logical value, character string or NULL value
specifying if and how zeros should be formatted specially.
Useful for pretty printing 'sparse' objects.}
    \item{\code{trim}}{a logical value; if \code{FALSE}, logical, numeric and complex values
are right-justified to a common width: if \code{TRUE} the leading blanks
for justification are suppressed.}
  }}
}
\value{
a character vector of length equal to the estimate vector.
}
\description{
Point and interval (pint)
}
\examples{

tbl_pint(1.12, 1.00, 1.25)

}
