% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblstring.R
\name{tbl_string}
\alias{tbl_string}
\title{Table strings}
\usage{
tbl_string(
  ...,
  .sep = "",
  .envir = parent.frame(),
  decimals_0_to_1 = 2,
  decimals_1_to_10 = 1,
  decimals_10_to_100 = 0,
  decimals_100_plus = 0,
  big_mark = ",",
  .missing = "--"
)
}
\arguments{
\item{...}{Strings to format. Multiple inputs are concatenated
together before formatting. Named arguments are not supported.}

\item{.sep}{Separator used to separate elements}

\item{.envir}{environment to evaluate each expression in.}

\item{decimals_0_to_1}{number of decimals to display for
numbers < 1}

\item{decimals_1_to_10}{number of decimals to display for
numbers >= 1 and < 10}

\item{decimals_10_to_100}{number of decimals to display for
numbers >= 10 and < 100}

\item{decimals_100_plus}{number of decimals to display for
numbers >= 100}

\item{big_mark}{a character value used to separate number groups to the
left of the decimal point. See \link{prettyNum} for more details on this.
Set this input to '' to negate it's effect.}

\item{.missing}{a character string that replaces missing values.}
}
\value{
a character vector
}
\description{
Table strings
}
\examples{

x <- runif(10)
y <- runif(10)

tbl_string("{x} / {y} = {x/y}")

tbl_string("{x}", "({100 * y}\%)", .sep = ' ')

df = data.frame(x = 1:10, y=1:10)

tbl_string("{x} / {y} = {as.integer(x/y)}", .envir = df)
tbl_string("{x} / {y} = {as.integer(x/y)}", .envir = NULL)

with(df, tbl_string("{x} / {y} = {as.integer(x/y)}"))

}
