% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbv_ci.R
\name{tbv_ci}
\alias{tbv_ci}
\title{Confidence intervals}
\usage{
tbv_ci(
  est,
  lwr,
  upr,
  fun = function(x) x,
  brac_left = "(",
  brac_mid = ", ",
  brac_right = ")"
)
}
\arguments{
\item{est}{numeric vector of point estimates.}

\item{lwr}{numeric vector of lower interval estimates.}

\item{upr}{numeric vector of upper interval estimates.}

\item{fun}{a function that will be applied to all numeric values before
they are multiplied by \code{mult_by}.
inputs before they are passed into \link{tbv_round}, but after \code{fun()}.}

\item{brac_left}{a character value that will close the left side of
the interval.}

\item{brac_mid}{a character value that will separate the lower and
upper portion of the intervals}

\item{brac_right}{a character value that will close the right side of
the interval.}
}
\value{
a character vector with each value comprising estimate and
interval values.
}
\description{
Confidence intervals
}
\examples{

tbv_ci(1.111, 0.555, 1.555)

}
