% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbv_pval.R
\name{tbv_pval}
\alias{tbv_pval}
\title{P-values}
\usage{
tbv_pval(x, threshold = 0.001)
}
\arguments{
\item{x}{a numeric vector, ideally one containing p-values.}

\item{threshold}{a numeric value indicating the lowest number that
should be displayed. Any numbers below this threshold will be
written as < threshold in the output.}
}
\value{
a character vector containing formatted p-values.
}
\description{
P-values
}
\note{
the number of decimal digits to display is determined based
on the threshold value. For example, if you set threshold as 0.01,
then only 2 digits are shown. Similarly, a threshold of 0.001 will
lead to displaying 3 digits.
}
\examples{

tbv_pval(c(0.1, 0.01, 0.001, 0.0001))
tbv_pval(c(0.1, 0.01, 0.001, 0.0001), threshold = 0.01)

}
