% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_gap.R
\name{as_pointGap}
\alias{as_pointGap}
\title{Point gap casting}
\usage{
as_pointGap(x)
}
\arguments{
\item{x}{object to be coerced. \code{x} can be a list with 3 elements or
a data.frame with 3 columns. The values of \code{x} should be ordered
as point, lower, and then upper values.}
}
\value{
point gap (\code{pointGap}) value(s).
}
\description{
Coerces object(s) into type \code{"pointGap"}.
}
\note{
data in \code{x} must be numeric. Valid objects are:
\itemize{
\item \code{vector} (length must = 3)
\item \code{matrix} (ncol must = 3)
\item \code{list} (length must = 3)
\item \code{data.frame} (ncol must = 3)
}

In each case,
\itemize{
\item the first value or column is taken as point values
\item the second value or column is taken as lower values
\item the third value or column is taken as upper values
}
}
\examples{

x <- list(c(1,2,3), c(0,1,2), c(2,3,4))

as_pointGap(x)

x <- as.data.frame(x)

as_pointGap(x)

x <- c(1L, 0L, 2L)

as_pointGap(x)

x <- matrix(c(1,2,3,0,1,2,2,3,4), ncol=3)

as_pointGap(x)


}
\seealso{
pointGap
}
