% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_gap.R
\name{format.tblStrings_pointGap}
\alias{format.tblStrings_pointGap}
\title{Print point gaps}
\usage{
\method{format}{tblStrings_pointGap}(x, ...)
}
\arguments{
\item{x}{an object of type \code{'pointGap'}}

\item{...}{not currently used}
}
\value{
printed output in the console window
}
\description{
Shows the values that will be printed in
tables or figures when a \code{pointGap} object is included.
}
\examples{

pointGap(1, 0, 2, max_decimal = 0)
pointGap(1, 0, 2, max_decimal = 1)
pointGap(1, 0, 2, max_decimal = 2)

pointGap(1e4, 1e3, 2e4)
pointGap(1e4, 1e3, 2e4, big_mark = '-')
pointGap(1e4, 1e3, 2e4, brac_left = '[', brac_right = ']', brac_middle = '-')

}
