% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_perc.R, R/point_error.R, R/point_gap.R
\name{is_numPer}
\alias{is_numPer}
\alias{is_pointErr}
\alias{is_pointGap}
\title{tblString inheritance}
\usage{
is_numPer(x)

is_pointErr(x)

is_pointGap(x)
}
\arguments{
\item{x}{an object}
}
\value{
a logical vector of length equal to \code{x}
}
\description{
tblString inheritance
}
\examples{

is_pointGap(1)

is_pointGap(pointGap(1,0,2))


}
