% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblstring.R
\name{tbl_string}
\alias{tbl_string}
\title{Table strings}
\usage{
tbl_string(..., .sep = "", .envir = parent.frame())
}
\arguments{
\item{...}{Strings to format. Multiple inputs are concatenated
together before formatting. Named arguments are not supported.}

\item{.sep}{Separator used to separate elements}

\item{.envir}{Environment to evaluate each expression in.}
}
\value{
Character valued vector
}
\description{
Table strings
}
\examples{

x <- runif(10)
y <- runif(10)

tbl_string("{x} / {y} = {x/y}")

tbl_string("{x}", "({100 * y}\%)", .sep = ' ')

df = data.frame(x = 1:10, y=1:10)

tbl_string("{x} / {y} = {as.integer(x/y)}", .envir = df)

with(df, tbl_string("{x} / {y} = {as.integer(x/y)}"))

}
