% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_string.R
\name{tbl_string}
\alias{tbl_string}
\title{Table strings}
\usage{
tbl_string(
  ...,
  .sep = "",
  .envir = parent.frame(),
  round_half_to_even = FALSE,
  breaks = c(1, 10, Inf),
  decimals = c(2, 1, 0),
  miss_replace = "--",
  big_mark = ",",
  big_interval = 3L,
  small_mark = "",
  small_interval = 5L,
  decimal_mark = getOption("OutDec"),
  zero_print = NULL,
  trim = TRUE
)
}
\arguments{
\item{...}{Strings to format. Multiple inputs are concatenated
together before formatting. Named arguments are not supported.}

\item{.sep}{Separator used to separate elements}

\item{.envir}{environment to evaluate each expression in.}

\item{round_half_to_even}{a logical value. If \code{TRUE}, then when the
fractional part of \code{x} is half-way between two rounding boundaries,
it will be rounded to the even boundary nearest to x. For example,
+23.5 becomes +24, as does +24.5; while −23.5 becomes −24,
as does −24.5. If \code{FALSE}, halves are automatically rounded to
the boundary with highest absolute value.}

\item{breaks}{a positive, monotonically increasing numeric vector
designating rounding boundaries. With \code{breaks = c(1, 10, 100, 1000)}
and \code{decimals = c(2, 1, 0)}}

\item{decimals}{a numeric vector of equal length to \code{breaks} that
indicates how many decimals to round to in the numeric range
designated by \code{breaks}. (see notes for example). Allowed values
are 0 <= \code{decimals} <= 20.}

\item{miss_replace}{a character string that replaces missing values.}

\item{big_mark}{a character value used to separate number groups to the
left of the decimal point. See \link{prettyNum} for more details on this.
Set this input to '' to negate it's effect.}

\item{big_interval}{a numeric value indicating the size of number groups
for numbers before (hence big) the decimal.}

\item{small_mark}{a character value used to separate number groups to
the right of the decimal point.}

\item{small_interval}{a numeric value indicating the size of number groups
for numbers after (hence small) the decimal.}

\item{decimal_mark}{the character to be used to indicate the numeric
decimal point.}

\item{zero_print}{a logical value, character string or NULL value
specifying if and how zeros should be formatted specially.
Useful for pretty printing 'sparse' objects.}

\item{trim}{a logical value; if \code{FALSE}, logical, numeric and complex values
are right-justified to a common width: if \code{TRUE} the leading blanks
for justification are suppressed.}
}
\value{
a character vector
}
\description{
Table strings
}
\examples{

x <- runif(10)
y <- runif(10)

tbl_string("{x} / {y} = {x/y}")

tbl_string("{x}", "({100 * y}\%)", .sep = ' ')

df = data.frame(x = 1:10, y=1:10)

tbl_string("{x} / {y} = {as.integer(x/y)}", .envir = df)
tbl_string("{x} / {y} = {as.integer(x/y)}", .envir = NULL)

with(df, tbl_string("{x} / {y} = {as.integer(x/y)}"))

}
