% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_val.R
\name{tbl_val}
\alias{tbl_val}
\title{Table value rounding}
\usage{
tbl_val(x, max_decimals = 2, big_mark = ",")
}
\arguments{
\item{x}{a vector of numeric values}

\item{max_decimals}{an integer value that will determine the maximum
number of decimals in the output. Larger numbers will not use the
maximum number of decimals in order to maintain the same, or at
least similar length as smaller numbers.}

\item{big_mark}{a character value used to separate number groups to the
left of the decimal point. See \link{prettyNum} for more details on this.
Set this input to '' to negate it's effect.}
}
\value{
a character vector with rounded values
}
\description{
Table value rounding
}
\note{
\code{NA} values are automatically converted to character \code{'NA'} values.
This is because different R packages that tabulate numbers handle
true \code{NA} values differently, but handle character values the same way.
}
\examples{

tbl_val( c(0.1234, 1.234, 12.34, 123.4, 1234) )

}
