% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblStrings-package.R
\docType{package}
\name{tblStrings-package}
\alias{tblStrings}
\alias{tblStrings-package}
\title{tblStrings: What the Package Does (One Line, Title Case)}
\description{
tblStrings helps translate double and integer valued data into
    character values formatted with rules that are generally consistent with
    the requested formats from academic journals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcjaeger/tblStrings}
  \item Report bugs at \url{https://github.com/bcjaeger/tblStrings/issues}
}

}
\author{
\strong{Maintainer}: Byron Jaeger \email{bcjaeger@uab.edu} (\href{https://orcid.org/0000-0001-7399-2299}{ORCID})

}
\keyword{internal}
